Create song database entity
